Setting up the unit tests to work with the interpreter
When running the tests with the instrumenter there is a risk of leaving the image in an unwanted state because of MetaLink installation.
By default we will run the tests with the AST interpreter.
When executing a full benchmark it is best to execute the parts related to the instrumenter at the end so that there is as little impact as possible on other parts.